+++ client/store/actionTypes.js
@@ -2,3 +2,6 @@ export const GET_PRODUCTS = 'GET_PRODUCTS'
 
 export const GET_ALL_REVIEWS = 'GET_ALL_REVIEWS'
 export const CREATE_REVIEW = 'CREATE_REVIEW'
+
+export const SET_AUTH = 'SET_AUTH'
+export const LOGOUT = 'LOGOUT'
