+++ client/store/thunks.js
@@ -1,7 +1,9 @@
 import axios from 'axios';
 
 import { _getProduct } from './actionCreators';
+import { _getAllReviews, _createReview } from './actionCreators';
 
+//PRODUCTS
 export const getProducts = () => {
   return (dispatch) => {
     return axios.get('/api/products')
@@ -11,6 +13,7 @@ export const getProducts = () => {
   }
 }
 
+//REVIEWS
 export const getAllReviews = (reviews) => {
   return (dispatch) => {
     axios.get('/api/reviews')
@@ -19,8 +22,6 @@ export const getAllReviews = (reviews) => {
       .catch(error => console.log(error.message))
   }
 }
-
-
 export const createReview = (review) => {
   return (dispatch) => {
     axios.post(`/api/products/${id}/reviews`, review)
@@ -28,4 +29,40 @@ export const createReview = (review) => {
       .then(review => dispatch(_createReview(review)))
       .catch(error => console.log(error.message))
   }
+}
+
+//AUTH
+export const exchangeTokenForAuth = history => {
+    return async dispatch => {
+        try {
+            const token = window.localStorage.getItem('token')
+            if(!token){
+                return
+            }
+            const user = await axios.get('/api/auth', { headers : {
+                authorization : token
+            }})
+            dispatch(_setAuth(user.data))
+            if(history){
+                history.push('/products')
+            }
+        } catch(err){
+            console.log(err)
+            window.localStorage.removeItem('token')
+        }
+    }
+}
+export const logIn = (credentials, history) => {
+    return async dispatch => {
+        const response = await axios.post('/api/auth/', credentials)
+        window.localStorage.setItem('token', response.data.token)
+        return dispatch(exchangeTokenForAuth(history))
+    }
+}
+export const logOut = history => {
+    return dispatch => {
+        window.localStorage.removeItem('token')
+        dispatch(_logOut())
+        history.push('/home')
+    }
 }
